moved convert to ConvertSpreadOperatorPass
Since we need to get the function array_merge_spread,
we need to do the conversion after the functions are collected.

Removed unnecessary tests.
