+++ compiler/pipes/convert-spread-operator.h
@@ -1,0 +1,19 @@
+// Compiler for PHP (aka KPHP)
+// Copyright (c) 2021 LLC «V Kontakte»
+// Distributed under the GPL v3 License, see LICENSE.notice.txt
+
+#pragma once
+
+#include "compiler/function-pass.h"
+
+// Converts the spread operator (...$a) to a call to the array_merge_spread function.
+class ConvertSpreadOperatorPass final : public FunctionPassBase {
+public:
+  string get_description() final {
+    return "Convert spread operator";
+  }
+
+  VertexPtr on_exit_vertex(VertexPtr root) override;
+
+  static VertexPtr convert_array_with_spread_operators(VertexAdaptor<op_array> array_vertex);
+};
