--- tests/phpt/variadic_args/104_unpacking_in_wrong_place_1.php
@@ -1,9 +1,0 @@
-@kphp_should_fail
-<?php
-
-function f(...$args) {
-    var_dump($args);
-}
-
-$a = [1, 2, 3];
-f(3 + ...$a);
