--- tests/phpt/variadic_args/105_unpacking_in_wrong_place_2.php
@@ -1,9 +1,0 @@
-@kphp_should_fail
-<?php
-
-function f($x, ...$args) {
-    var_dump($args);
-}
-
-$a = [1, 2, 3];
-f(10, ...$a + 3);
